gha: additionally test kvstore mode in E2E upgrade workflow
Modify a few matrix entries of the Cilium E2E Upgrade workflow to
configure Cilium in kvstore mode, to cover this functionality here
as well in addition to the clustermesh workflows. In detail, the
etcd instance is executed as a pod running in host network, which
is setup via the `kind-kvstore-start` makefile target. The matrix
entries are selected trying to cover the most common combinations,
that is native-routing/tunneling, KPR off/on and wireguard off/on,
and avoiding incompatible options (mainly Egress Gateway).

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>