gha: additionally test kvstore mode in IPSec workflows
Modify a few matrix entries of the Conformance IPSec and IPSec Upgrade
workflows to configure Cilium in kvstore mode, to cover this functionality
here as well in addition to the E2E and clustermesh workflows. In detail,
the etcd instance is executed as a pod running in host network, which
is setup via the `kind-kvstore-start` makefile target. The matrix entries
are selected to cover both native routing and tunneling, while avoiding
incompatible options (mainly Egress Gateway and Mutual Auth).

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>