daemon: Reduce level of socket LB tracing warning
This warning is emitted when the kernel doesn't support socket-level
LB tracing. That feature is however enabled by default and only kernels
5.7+ support it. That means this warning will be displayed by default to
a lot of users.

This commit reduces the log level to Info to avoid that and to be
consistent with other advanced features that are enabled by default.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>