operator: fix race condition when deleting nodes from kvstore
The operator already includes a reconciliation logic to delete node
entries from the kvstore when the corresponding CiliumNode gets deleted.

However, it is currently affected by a race condition, because the
CiliumNode gets typically deleted as a consequence of the Node object
being deleted (as they are bound via an owner reference), although at
that point the Cilium agent on that node may still be running. If
that's the case, it would detect the removal of the kvstore entry,
and react recreating it. Hence, defeating the whole purpose of the
operator logic, and leading to the node entry being eventually deleted
by the lease expiration only. The occurrence of this race condition
is typically signaled by the "Received delete event for key which
re-appeared within delay time window" log message.

Let's address this by extending the GC logic to wait deleting the
node entry until the Cilium agent running on that node has been
stopped, to make sure it will not be recreated right away. The rate
limiter of the associated working queue is also customized to
increase the base delay (5ms by default), considering that it is
pointless to retry with a high frequency in this case.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>